refactor: Application 클래스 코드 RaceGame으로 이동
